refactor(WindowUtils): Move `read_window` in with the window writer functions
